requirements.txt
attrs==23.1.0
blinker==1.6.3
click==8.1.7
exceptiongroup==1.1.3
flasgger==0.9.7.1
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.2
mysqlclient==2.2.0
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
SQLAlchemy==2.0.22
tomli==2.0.1
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
